refactor task project manager ui to delete options and add them to UI messages
