+++ src/main/kotlin/di/UiModule.kt
@@ -10,7 +10,6 @@ import org.example.ui.features.state.AdminEntityStateManagerUi
 import org.example.ui.features.state.EntityStateManagerUi
 import org.example.ui.features.state.ShowAllEntityStateManagerUi
 import org.example.ui.features.task.TaskManagerUi
-import org.example.ui.features.task.TaskManagerUiImp
 import org.example.ui.features.user.CreateUserUi
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
@@ -25,7 +24,7 @@ val uiModule = module {
     factory<AddUserToProjectUI> { AddUserToProjectUI(get(), get(), get(), get(), get()) }
     factory<ShowAllEntityStateManagerUi> { ShowAllEntityStateManagerUi(get(), get()) }
     factory<AdminEntityStateManagerUi> { AdminEntityStateManagerUi(get(), get(), get(), get(), get()) }
-    factory<TaskManagerUi> { TaskManagerUiImp(get(), get(), get(), get(), get(), get(), get(), get(), get()) }
+    factory<TaskManagerUi> { TaskManagerUi(get(), get(), get(), get(), get(), get(), get(), get(), get()) }
     factory<EntityStateManagerUi> { EntityStateManagerUi(get(), get(), get(), get()) }
     factory<UiUtils> { UiUtils() }
 }
