+++ src/main/kotlin/ui/PlanMateConsoleUi.kt
@@ -8,7 +8,6 @@ import org.example.ui.features.common.utils.UiMessages
 import org.example.ui.features.login.LoginUi
 import org.example.ui.features.project.ProjectManagerUi
 import org.example.ui.features.state.EntityStateManagerUi
-import org.example.ui.features.task.TaskManagerUi
 import org.example.ui.features.user.CreateUserUi
 import org.example.ui.input_output.input.InputReader
 import org.example.ui.input_output.output.OutputPrinter
