+++ src/main/kotlin/ui/features/common/utils/UiMessages.kt
@@ -91,19 +91,20 @@ object UiMessages {
 
     //audit constants
     val SHOW_AUDIT_OPTIONS = """
-        |--------------- Audit Menu --------------|
+        |=============== Audit Menu ==============|
         | 1. Show Project Audits                  |
         | 2. Show Task Audits                     |
         | 3. Show All Audits                      |
         | 0. Exit                                 |
+        |-----------------------------------------|
     """.trimIndent()
     const val PROMPT_TASK_NAME = "Enter Task name: "
     const val PROMPT_PROJECT_NAME = "Enter Project Name: "
 
 
     val MAIN_MENU_WELCOME_MESSAGE_FOR_ADMIN = """
-        |----------- Welcome to planMate System ^_^ -----------|
-        |-------  Please choose what do you want to do  -------|
+        |=========== Welcome to planMate System ^_^ ===========|
+        |=======  Please choose what do you want to do  =======|
         | 1. Mange projects                                    |
         | 2. Manage tasks                                      |
         | 3. Manage states                                     |
@@ -115,8 +116,8 @@ object UiMessages {
     """.trimIndent()
 
     val MAIN_MENU_WELCOME_MESSAGE_FOR_MATE = """
-        |----------- Welcome to planMate System ^_^ -----------|
-        |-------  Please choose what do you want to do  -------|
+        |=========== Welcome to planMate System ^_^ ===========|
+        |=======  Please choose what do you want to do  =======|
         | 1. Manage tasks                                      |
         | 2. Manage states                                     |
         | 3. Show Audit                                        |
@@ -126,7 +127,7 @@ object UiMessages {
 
 
     val SHOW_ADD_USER_TO_PROJECT_OPTIONS = """
-        |--------------- Users In Project Management --------------|
+        |=============== Users In Project Management ==============|
         | 1. Assign users to project                               |
         | 2. View users assigned to project                        |
         | 3. Remove user from project                              |
@@ -135,7 +136,7 @@ object UiMessages {
     """.trimIndent()
 
     val SHOW_ADMIN_ENTITY_STATE_OPTIONS = """
-        |----------------- Entity State Management ----------------|
+        |=============== Entity State Management ==================|
         | 1. Show all states                                       |
         | 2. Add state                                             |
         | 3. Update state                                          |
@@ -145,12 +146,25 @@ object UiMessages {
     """.trimIndent()
 
     val SHOW_PROJECT_MANAGEMENT_OPTIONS = """
-        |------------------ Project Management --------------------|
+        |=================== Project Management ===================|
         | 1. Show all projects                                     |
         | 2. Show project details                                  |
         | 3. Add project                                           |
         | 4. Update project                                        |
         | 5. Delete project                                        |
+        | 0. Exit                                                  |
+        |----------------------------------------------------------|
+    """.trimIndent()
+
+    val SHOW_TASK_MANAGEMENT_OPTIONS = """
+        |==================== Tasks Management ====================|
+        | 1. Show all tasks                                        |
+        | 2. Show task by name                                     |
+        | 3. Add task                                              |
+        | 4. Update task                                           |
+        | 5. Delete task                                           |
+        | 6. Show tasks by project name                            |
+        | 7. Show mate task assignments                            |
         | 0. Exit                                                  |
         |----------------------------------------------------------|
     """.trimIndent()
